refactor prompts page into subcomponents
