+++ src/components/prompts/demoData.ts
@@ -1,0 +1,20 @@
+export const colorTokens = [
+  "bg-ring",
+  "bg-accent",
+  "bg-accent-2",
+  "bg-lavDeep",
+  "bg-danger",
+  "bg-success",
+  "bg-glow",
+];
+
+export const spacingTokens = [4, 8, 12, 16, 24, 32, 48, 64];
+
+export const glowTokens = ["--glow-strong", "--glow-soft"];
+
+export const focusRingToken = "--theme-ring";
+
+export const radiusTokens = ["--radius-md", "--radius-lg", "--radius-xl", "--radius-2xl"];
+export const radiusClasses = ["rounded-md", "rounded-lg", "rounded-xl", "rounded-2xl"];
+
+export const typeRamp = ["eyebrow", "title", "subtitle", "body", "caption"];
